+++ intent_server/dataset.py
@@ -18,6 +18,9 @@ class Dataset:
         output['values'] = self.data.T.to_dict()
         return output
 
+    def labels(self) -> pd.DataFrame:
+        return self.data[self.label]
+
     def numerical(self) -> pd.DataFrame:
         return self.data.select_dtypes(include='number')
